pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devdash"
version = "0.1.0"
description = "Open-source macOS menubar developer utilities"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "DevDash Contributors"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "rumps>=0.4.0",
    "pyperclip>=1.8.0",
    "PyJWT>=2.8.0",
    "pyyaml>=6.0",
    "croniter>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "pre-commit",
    "build",
    "twine",
]

[project.scripts]
devdash = "devdash.__main__:main"

[project.urls]
Homepage = "https://github.com/devdash/devdash"
Repository = "https://github.com/devdash/devdash"
Issues = "https://github.com/devdash/devdash/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/devdash --cov-report=term-missing"
